+++ src/MainWindow.xaml.cs
@@ -58,11 +58,11 @@ namespace NinjaCalc {
             //========== ELECTRONICS ===========//
 
             // BASIC
-            this.RegisterCalculator(new Calculators.Electronics.Basic.OhmsLaw.OhmsLawCalculator());
-            this.RegisterCalculator(new Calculators.Electronics.Basic.ResistorDivider.ResistorDividerCalc());
+            //this.RegisterCalculator(new Calculators.Electronics.Basic.OhmsLaw.OhmsLawCalculator());
+            //this.RegisterCalculator(new Calculators.Electronics.Basic.ResistorDivider.ResistorDividerCalc());
 
             // PCB
-            this.RegisterCalculator(new Calculators.Electronics.Pcb.TrackCurrentIpc2221A.TrackCurrentIpc2221ACalculator());
+            //this.RegisterCalculator(new Calculators.Electronics.Pcb.TrackCurrentIpc2221A.TrackCurrentIpc2221ACalculator());
             this.RegisterCalculator(new Calculators.Electronics.Pcb.TrackCurrentIpc2152.TrackCurrentIpc2152Calculator());
             
         }